feat: Add cancel button
