+++ packages/app_center/lib/widgets/widgets.dart
@@ -1,3 +1,4 @@
+export 'active_change_content.dart';
 export 'app_card.dart';
 export 'app_icon.dart';
 export 'app_info_bar.dart';
