+++ packages/app_center_ratings_client/lib/src/ratings.freezed.dart
@@ -20,9 +20,7 @@ mixin _$Rating {
   int get totalVotes => throw _privateConstructorUsedError;
   RatingsBand get ratingsBand => throw _privateConstructorUsedError;
 
-  /// Create a copy of Rating
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   $RatingCopyWith<Rating> get copyWith => throw _privateConstructorUsedError;
 }
 
@@ -44,8 +42,6 @@ class _$RatingCopyWithImpl<$Res, $Val extends Rating>
   // ignore: unused_field
   final $Res Function($Val) _then;
 
-  /// Create a copy of Rating
-  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -88,8 +84,6 @@ class __$$RatingImplCopyWithImpl<$Res>
       _$RatingImpl _value, $Res Function(_$RatingImpl) _then)
       : super(_value, _then);
 
-  /// Create a copy of Rating
-  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -150,9 +144,7 @@ class _$RatingImpl implements _Rating {
   @override
   int get hashCode => Object.hash(runtimeType, snapId, totalVotes, ratingsBand);
 
-  /// Create a copy of Rating
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   @override
   @pragma('vm:prefer-inline')
   _$$RatingImplCopyWith<_$RatingImpl> get copyWith =>
@@ -171,11 +163,8 @@ abstract class _Rating implements Rating {
   int get totalVotes;
   @override
   RatingsBand get ratingsBand;
-
-  /// Create a copy of Rating
-  /// with the given fields replaced by the non-null parameter values.
   @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   _$$RatingImplCopyWith<_$RatingImpl> get copyWith =>
       throw _privateConstructorUsedError;
 }
